The DTSSignals Module

DTSSignals — Module

DTSSignals (version 0.0.2) --- An in-house module for signals processing in DTS.

Author: FVA, all rights reserved

Module Index

DTSSignals.Signal

DTSSignals.convolve2

DTSSignals.sample

DTSSignals.shift

DTSSignals.sinusoid

DTSSignals.trim

DTSSignals.visualise!

DTSSignals.δ

Detailed API

DTSSignals.Signal — Type

struct Signal
    n::Vector{Int}
    v::Vector{Complex}#vector of values
    fs::Float64#sampling frequency
end

A type to encode signals, e.g. two co-indexed sequences of integer time indices [n] and values [v]. Note that these two sequences must have the same length, and this is enforced by the constructor.

If we want to ground the signal in some sampling process, we can add a sampling frequency [fs]. If ommitted, we leave undefined, so cave canem.

Base.:* — Method

scale(α::Complex,x::signal) -> Signal α

A function to scale a signal x by a scalar α. Scalars can be ComplexF64, Float64 or even Note that the scalar outer product can be abbreviated to nothing so we can elide it.

using Test
@test scale(Complex(2.0),δ(0)) == δ(0) + δ(0)
@test (2.0+0im)*δ(0) == δ(0) + δ(0)#Not yet defined.
@test 2δ(0) == δ(0) + δ(0) == 2.0δ(0) == (2.0+0.0im) * δ(0)

Base.:* — Method

*(x::Signal, y::Signal) → Complex

Dot-product of two signal. In the spirit of extending the product.

xs = δ(0) + δ(4) + δ(5)
ys = δ(0) + 2δ(4) 
@test x * y == 4.0

Base.:+ — Method

add(x::Signal, y::Signal) → Signal
x + y -> signal

A function to add two signals.

Example:

x = δ(0) 
y = δ(5)
w = x + y
x = add(δ(4), w)
using Test
@test w == δ(0) + δ(5)

Base.:== — Method

==(x::Signal, y::Signal) → Bool

An equality predicate for signals. We test for equality of all content.

Base.conj — Method

conj(x::Signal) → Signal

A function to generate the conjugate of a signal.

s = δ(0)
r = conj(s)
using Test
@test !isreal(s.v) || x == r#only true for real signals.
@test s == conj(r)## Always true since conj is an involution.
@test (π/4)s != conj((π/4)s)

Base.imag — Method

imag(x::Signal) → Signal

A function to obtain the imaginary part of a signal.

Base.isreal — Method

isreal(x::Signal) → Bool

A predicate to check if a signal is real.

using Test
@test isreal(Signal([1,2,3],[1.0,2.0,3.0])) && isreal(δ(0))
@test !isreal(Signal([1,2,3],[1.0,2.0,3.0im]))

Base.length — Method

length(x::Signal) → Integer

A primitive to observe the lenth of a signal

Base.real — Method

real(x::Signal) → Signal

A function to obtain the real part of a signal.

Base.zero — Method

zero(Signal) → Signal

The additive zero of signals. It is also the convolutive zero.

DSP.conv — Method

  convolve(x::Signal, y::Signal)::Signal

A primitive to convolve two signals. Example:

x = δ(2)
y = δ(5)
@assert conv(x,y) == δ(7)

DTSSignals.convolve2 — Method

Another implementation, using matrix multiplication, unexported.

DTSSignals.sample — Method

sample(f::Function, fs::AbstractFloat, N::Integer) → Signal

A function to sample a time function [f] at a given sampling frequency [fs] for [N] samples.

using Test
x = sample(t -> sin(2π*t), 1.0, 10)
@test length(x.n) == 10
@test x.v  == collect(0:9)
@test x.v ≈ sin.(2π*collect(0:9))

DTSSignals.shift — Method

shift(x::signal, τ::Integer ) -> signal

A function to implement a time displacement (positive delay).

Example:

@assert shift(δ(0), 4) == δ(4)

DTSSignals.sinusoid — Method

sinusoid(A, f, φ, t) → Real

A function to generate a real sinusoid with its standard parameters:

A: amplitude, dimensionless.

f: frequency in Hz.

φ: phase, in radians.

t: time in seconds.

Note that the sinusoid is generated with a cosine function.

DTSSignals.trim — Method

trim(t, v) → (trimmedt,trimmedv)

A function to trim away the zeros of a value vector on the left and the right, and co-indexed time scale. This allows some sparsity in the representation of signals with a finite number of non-zero values.

Use example:

n = collect(-3:3);
v = Vector{Float64}zeros(length(t)); 
v[2] = 1.0;#This is the value of a delta at n=-2, δ(n + 2)
(trimmedt, trimmedv) = trim(n,v)

DTSSignals.visualise! — Method

visualise!(x::signal)

A primitive to visualise a DT signal as a stem plot:

The default is to visualize the real component

If "full" visualization is requested the default is polar

For real vs. imaginary part use full=true, polar=false

n = collect(-3:3)
v = Complex.([0.0, 0.0, 1.0, 1.0, 1.0, 0.0, 0.0])
#v = map(Complex, [0.0, 1.0, 1.0, 1.0, 1.0, 0.0, 0.0])
x = Signal(n,v; doTrim = false)#The N=3 pulse
f = Figure()
#Axis(f[1,1])
visualise!(x)
Axis(f[end+1,1])
visualise!(x;full=true)
Axis(f[end+1,1])
visualise!(x;full=true,polar=false)
f

DTSSignals.δ — Method

δ(τ::Integer) → Signal

A function to generate a delta signal at a shift of τ.

Example:

δ(5)